Split Asset hierarchy and added workers
