+++ src/models/playerAsset/Army.java
@@ -9,7 +9,9 @@ import java.util.Arrays;
   Battle group units have met at the Rally Point and are able to move together and fight
   Reinforcements are units on the way to a Rally Point
  */
-public class Army extends PlayerAsset {
+//TODO: Add Army Iterators
+
+public class Army extends CombatAsset {
     private ArrayList<Unit> battleGroup = new ArrayList<>();
     private ArrayList<Unit> reinforcements = new ArrayList<>();
 
@@ -45,7 +47,6 @@ public class Army extends PlayerAsset {
         return battleGroup;
     }
 
-    @Override
     public void accept(AssetVisitor v) {
         v.visitArmy(this);
         for (Unit u : reinforcements){
