+++ src/models/playerAsset/NonCombatAsset.java
@@ -1,0 +1,4 @@
+package models.playerAsset;
+
+public class NonCombatAsset {
+}
