+++ src/models/playerAsset/Player.java
@@ -35,6 +35,8 @@ public class Player {
         units.accept(v);
         structures.accept(v);
     }
+
+
     public AssetIterator<PlayerAsset, TypeIterator<PlayerAsset, Iterator<PlayerAsset>>> makeAssetIterator(ArrayList<TypeIterator<PlayerAsset, Iterator<PlayerAsset>>> list) {
         return new AssetIterator<PlayerAsset, TypeIterator<PlayerAsset, Iterator<PlayerAsset>>>(){
