+++ src/models/playerAsset/RallyPoint.java
@@ -2,7 +2,7 @@ package models.playerAsset;
 
 import models.visitor.AssetVisitor;
 
-public class RallyPoint extends PlayerAsset{
+public class RallyPoint extends NonCombatAsset{
     private Army army;
 
     public void accept(AssetVisitor v){
