+++ src/models/playerAsset/Unit.java
@@ -5,7 +5,7 @@ import models.visitor.AssetVisitor;
 /* High level unit class used to define the Unit type
 	and define some basic functionality
  */
-public abstract class Unit extends PlayerAsset{
+public abstract class Unit extends CombatAsset{
 	
     private double movesPerTurn; 	//REVISIT AND MODIFY
 
@@ -17,7 +17,6 @@ public abstract class Unit extends PlayerAsset{
 	public double getMovesPerTurn(){return movesPerTurn;}
 	public void setMovesPerTurn(double m){this.movesPerTurn = m;}
 
-    @Override
     public void accept(AssetVisitor v) {
         v.visitUnit(this);
     }
