+++ src/models/playerAsset/Worker.java
@@ -1,0 +1,12 @@
+package models.playerAsset;
+
+public class Worker {
+    int production;
+
+    public Worker(){
+        production = 1;
+    }
+
+    public String getType(){ return "Worker";}
+    public void setProduction(int i){ production = i;}
+}
